backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic-settings==2.0.3
python-decouple==3.8
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
celery==5.3.4
redis==5.0.1
pandas==2.1.3
plotly==5.17.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.7
